Bounded World, Drawable World.
Added a BoundedWorld class, that has boundries. It also has a generateImage
method, that generates a drawable representation of the world. The GameObject
class now has a private variable that can contain it's image, as well as
set/get methods for it. Due to the visibility aspect of the GameObject, the
image is not a parameter of the constructor, and must be set seperately.

Also, added the Slick libraries, for implimentation purposes of the
graphics methods.

Signed-off-by: DanSze <danzap95@gmail.com>